test: SystemMessageTabComponent 테스트 코드 -> TestsPageTests.cs로 이동
